Only use DoApp to assign local or remote runners for refresh
This commit introduces a new CLI option, WithOptionalApp, which allows
CLIs like `status` to optionally be initialized with a config and a
project client without a hard requirement on an app existing. It now
uses DoApp directly to run a status report job. If the project is
remote, and the CLI is not inside the project dir, it sets the ref app
for the project client to use when queueing the status report job.